+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Presenters can now be specified using the `only:` keyword
   argument, which takes either a Class or an Array of classes
 - Replace repeated method default arguments with splats
+- Add assertions to TestHelper
 
 0.1.2
 -----